Fixing simple types init in indexer and long tests (changing 'long' to 'long long' for now)

+++ tests/testData/integration/arguments/long.h
@@ -1,5 +1,5 @@
 #import <Foundation/NSObject.h>
 
 @interface A : NSObject
-+ (void) foo: (long) arg;
++ (void) foo: (long long) arg;
 @end
